refactor(clawdstrike-policy-event): extract kinds/response_{request,shared,execution,rollback,acknowledgement} from edr/receipt/mod.rs
