+++ crates/libs/clawdstrike-policy-event/src/edr/receipt/kinds/mod.rs
@@ -10,6 +10,11 @@ mod policy_delta;
 mod policy_event;
 mod privacy;
 mod provider_degradation;
+mod response_acknowledgement;
+mod response_execution;
+mod response_request;
+mod response_rollback;
+mod response_shared;
 mod sensor_state;
 mod simulation;
 
@@ -25,5 +30,10 @@ pub(crate) use policy_delta::*;
 pub(crate) use policy_event::*;
 pub(crate) use privacy::*;
 pub(crate) use provider_degradation::*;
+pub(crate) use response_acknowledgement::*;
+pub(crate) use response_execution::*;
+pub(crate) use response_request::*;
+pub(crate) use response_rollback::*;
+pub(crate) use response_shared::*;
 pub(crate) use sensor_state::*;
 pub(crate) use simulation::*;
