+++ crates/libs/clawdstrike-policy-event/src/edr/receipt/mod.rs
@@ -2127,1401 +2127,3 @@ impl EndpointDecisionReceipt {
         Ok(())
     }
 }
-
-fn require_response_request_receipt_evidence(
-    decision: &EndpointDecisionRecord,
-    evidence: &[EndpointReceiptEvidence],
-    root_node_id: &str,
-    graph_slice_id: &str,
-    graph_content_hash: Option<&str>,
-    ttl_seconds: u64,
-    rollback_ref: &str,
-) -> Result<()> {
-    let dry_run = response_request_dry_run_from_decision(decision)?;
-    let response_action_id = decision
-        .finding_id
-        .as_deref()
-        .ok_or_else(|| anyhow!("response action id is required"))?;
-    require_response_action_id_matches_signed_response_fields(
-        response_action_id,
-        root_node_id,
-        graph_slice_id,
-        &decision.action,
-        ttl_seconds,
-        if dry_run { "dry_run" } else { "execute" },
-    )?;
-    let expected_rollback_ref =
-        expected_response_rollback_ref(&decision.action, dry_run, response_action_id);
-    if rollback_ref != expected_rollback_ref {
-        return Err(anyhow!(
-            "response rollback evidence hash must match signed response action fields"
-        ));
-    }
-    require_response_receipt_evidence_fields(
-        evidence,
-        response_action_id,
-        root_node_id,
-        graph_slice_id,
-        graph_content_hash,
-        ttl_seconds,
-        rollback_ref,
-    )
-}
-
-fn require_response_receipt_evidence(
-    evidence: &[EndpointReceiptEvidence],
-    action: &EndpointDecisionAction,
-    root_node_id: &str,
-    graph_slice_id: &str,
-    ttl_seconds: u64,
-    rollback_ref: &str,
-) -> Result<()> {
-    let response_action_id = response_action_id_from_rollback_ref(action, rollback_ref)?;
-    require_response_action_id_matches_signed_response_fields(
-        response_action_id.as_str(),
-        root_node_id,
-        graph_slice_id,
-        action,
-        ttl_seconds,
-        "execute",
-    )?;
-    require_evidence_value_hash(
-        evidence,
-        "responseActionId",
-        response_action_id.as_str(),
-        "response action id evidence",
-    )?;
-    let expected_rollback_ref = expected_live_response_rollback_ref(action, &response_action_id);
-    if rollback_ref != expected_rollback_ref {
-        return Err(anyhow!(
-            "response rollback evidence hash must match signed response action fields"
-        ));
-    }
-    require_response_receipt_evidence_fields(
-        evidence,
-        response_action_id.as_str(),
-        root_node_id,
-        graph_slice_id,
-        None,
-        ttl_seconds,
-        rollback_ref,
-    )
-}
-
-fn require_response_receipt_evidence_fields(
-    evidence: &[EndpointReceiptEvidence],
-    response_action_id: &str,
-    root_node_id: &str,
-    graph_slice_id: &str,
-    graph_content_hash: Option<&str>,
-    ttl_seconds: u64,
-    rollback_ref: &str,
-) -> Result<()> {
-    require_evidence_value_hash(
-        evidence,
-        "responseActionId",
-        response_action_id,
-        "response action id evidence",
-    )?;
-    require_evidence_value_hash(
-        evidence,
-        "rootNodeId",
-        root_node_id,
-        "response root node evidence",
-    )?;
-    require_evidence_value_hash(
-        evidence,
-        "graphSliceId",
-        graph_slice_id,
-        "response graph slice evidence",
-    )?;
-    if let Some(graph_content_hash) = graph_content_hash {
-        require_evidence_value_hash(
-            evidence,
-            "contentHash",
-            graph_content_hash,
-            "response graph content hash evidence",
-        )?;
-    }
-    require_evidence_value_hash(
-        evidence,
-        "ttlSeconds",
-        ttl_seconds.to_string(),
-        "response ttl evidence",
-    )?;
-    require_evidence_value_hash(
-        evidence,
-        "rollbackRef",
-        rollback_ref,
-        "response rollback evidence",
-    )?;
-    Ok(())
-}
-
-fn response_action_id_from_rollback_ref(
-    action: &EndpointDecisionAction,
-    rollback_ref: &str,
-) -> Result<String> {
-    let action_id = if action == &EndpointDecisionAction::CollectEvidence {
-        rollback_ref
-            .strip_prefix("rollback:noop:")
-            .or_else(|| rollback_ref.strip_prefix("rollback:"))
-    } else {
-        rollback_ref.strip_prefix("rollback:")
-    }
-    .ok_or_else(|| {
-        anyhow!("response rollback evidence hash must match signed response action fields")
-    })?;
-    require_nonempty(action_id, "response action id")?;
-    Ok(action_id.to_string())
-}
-
-fn require_response_action_id_matches_signed_response_fields(
-    response_action_id: &str,
-    root_node_id: &str,
-    graph_slice_id: &str,
-    action: &EndpointDecisionAction,
-    ttl_seconds: u64,
-    mode: &str,
-) -> Result<()> {
-    let stable_response_action_id = response_action_id_from_signed_response_fields_with_mode(
-        root_node_id,
-        graph_slice_id,
-        action,
-        ttl_seconds,
-        mode,
-    );
-    if response_action_id == stable_response_action_id {
-        return Ok(());
-    }
-    let Some(issuance_id) = response_action_id
-        .strip_prefix(stable_response_action_id.as_str())
-        .and_then(|suffix| suffix.strip_prefix(':'))
-    else {
-        return Err(anyhow!(
-            "response action id evidence hash must match signed response action fields"
-        ));
-    };
-    Uuid::parse_str(issuance_id).map_err(|_| {
-        anyhow!("response action id evidence hash must match signed response action fields")
-    })?;
-    Ok(())
-}
-
-#[cfg(test)]
-pub(crate) fn response_action_id_from_signed_response_fields(
-    root_node_id: &str,
-    graph_slice_id: &str,
-    action: &EndpointDecisionAction,
-    ttl_seconds: u64,
-) -> String {
-    response_action_id_from_signed_response_fields_with_mode(
-        root_node_id,
-        graph_slice_id,
-        action,
-        ttl_seconds,
-        "execute",
-    )
-}
-
-fn response_action_id_from_signed_response_fields_with_mode(
-    root_node_id: &str,
-    graph_slice_id: &str,
-    action: &EndpointDecisionAction,
-    ttl_seconds: u64,
-    mode: &str,
-) -> String {
-    let ttl = ttl_seconds.to_string();
-    stable_id(
-        "response_action",
-        [
-            root_node_id,
-            graph_slice_id,
-            action.as_str(),
-            mode,
-            ttl.as_str(),
-        ],
-    )
-}
-
-fn expected_response_rollback_ref(
-    action: &EndpointDecisionAction,
-    dry_run: bool,
-    response_action_id: &str,
-) -> String {
-    if !dry_run && action == &EndpointDecisionAction::CollectEvidence {
-        format!("rollback:noop:{response_action_id}")
-    } else {
-        format!("rollback:{response_action_id}")
-    }
-}
-
-fn expected_live_response_rollback_ref(
-    action: &EndpointDecisionAction,
-    response_action_id: &str,
-) -> String {
-    expected_response_rollback_ref(action, false, response_action_id)
-}
-
-fn require_response_family_id_evidence(
-    evidence: &[EndpointReceiptEvidence],
-    key: &str,
-    signed_id: Option<&str>,
-    field_name: &str,
-) -> Result<()> {
-    let signed_id = signed_id.ok_or_else(|| anyhow!("{field_name} signed id is required"))?;
-    require_evidence_value_hash(evidence, key, signed_id, field_name)
-}
-
-fn require_response_request_dry_run_evidence(
-    decision: &EndpointDecisionRecord,
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<()> {
-    let expected_dry_run = if response_request_dry_run_from_decision(decision)? {
-        "true"
-    } else {
-        "false"
-    };
-    require_evidence_value_hash(
-        evidence,
-        "dryRun",
-        expected_dry_run,
-        "response dry-run evidence",
-    )
-}
-
-fn response_request_dry_run_from_decision(decision: &EndpointDecisionRecord) -> Result<bool> {
-    let title = decision
-        .title
-        .as_deref()
-        .ok_or_else(|| anyhow!("response request title is required"))?;
-    match title {
-        "Endpoint response action dry run planned" => Ok(true),
-        "Endpoint response action planned" => Ok(false),
-        _ => Err(anyhow!("response request title is invalid")),
-    }
-}
-
-fn require_response_reason_evidence(evidence: &[EndpointReceiptEvidence]) -> Result<()> {
-    let Some(item) = evidence.iter().find(|item| item.key == "reason") else {
-        return Err(anyhow!("response reason evidence is required"));
-    };
-    let empty_reason_hash = sha256(b"").to_hex_prefixed();
-    if item.value_hash == empty_reason_hash {
-        return Err(anyhow!("response reason evidence must not be empty"));
-    }
-    Ok(())
-}
-
-fn require_response_execution_id_evidence(
-    evidence: &[EndpointReceiptEvidence],
-    decision: &EndpointDecisionRecord,
-    graph_slice_id: &str,
-    graph_content_hash: Option<&str>,
-) -> Result<()> {
-    require_response_family_id_evidence(
-        evidence,
-        "executionId",
-        decision.finding_id.as_deref(),
-        "execution id evidence",
-    )?;
-    let Some(expected_execution_id) = response_execution_id_from_signed_fields(
-        decision,
-        graph_slice_id,
-        graph_content_hash,
-        evidence,
-    )?
-    else {
-        return Ok(());
-    };
-    if decision.finding_id.as_deref() != Some(expected_execution_id.as_str()) {
-        return Err(anyhow!(
-            "execution id evidence hash must match signed response action fields"
-        ));
-    }
-    require_evidence_value_hash(
-        evidence,
-        "executionId",
-        expected_execution_id.as_str(),
-        "execution id evidence",
-    )
-}
-
-fn require_response_execution_evidence_bundle_evidence(
-    evidence: &[EndpointReceiptEvidence],
-    decision: &EndpointDecisionRecord,
-    graph_slice_id: &str,
-    graph_content_hash: Option<&str>,
-) -> Result<()> {
-    let graph_content_hash = graph_content_hash
-        .ok_or_else(|| anyhow!("execution evidence bundle graph content hash is required"))?;
-    let evidence_bundle_id = response_execution_bundle_id_from_signed_fields(
-        decision,
-        graph_slice_id,
-        graph_content_hash,
-    )?;
-    require_evidence_value_hash(
-        evidence,
-        "evidenceBundleId",
-        evidence_bundle_id.as_str(),
-        "execution evidence bundle id evidence",
-    )?;
-    require_evidence_value_hash(
-        evidence,
-        "evidenceBundleContentHash",
-        graph_content_hash,
-        "execution evidence bundle content hash evidence",
-    )
-}
-
-fn response_execution_id_from_signed_fields(
-    decision: &EndpointDecisionRecord,
-    graph_slice_id: &str,
-    graph_content_hash: Option<&str>,
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<Option<String>> {
-    let status = response_execution_status_from_decision(decision)?;
-    let graph_content_hash =
-        graph_content_hash.ok_or_else(|| anyhow!("execution id graph content hash is required"))?;
-    let evidence_bundle_id = response_execution_bundle_id_from_signed_fields(
-        decision,
-        graph_slice_id,
-        graph_content_hash,
-    )?;
-    let rollback_ref = decision
-        .rollback_ref
-        .as_deref()
-        .ok_or_else(|| anyhow!("response rollback ref is required"))?;
-    let response_action_id = response_action_id_from_rollback_ref(&decision.action, rollback_ref)?;
-    if let Some(prefix) = response_execution_transition_id_prefix(status) {
-        let reason_hash = response_execution_reason_evidence_hash(evidence)?;
-        return Ok(Some(response_execution_transition_id_from_reason_hash(
-            prefix,
-            response_action_id.as_str(),
-            evidence_bundle_id.as_str(),
-            rollback_ref,
-            reason_hash,
-        )));
-    }
-    if let Some(effect_binding_digest) =
-        response_execution_effect_binding_digest_from_evidence(evidence)?
-    {
-        if decision.action == EndpointDecisionAction::CollectEvidence {
-            return Err(anyhow!(
-                "collect evidence execution effect evidence is invalid"
-            ));
-        }
-        return Ok(Some(response_execution_id_from_effect_digest(
-            response_action_id.as_str(),
-            evidence_bundle_id.as_str(),
-            effect_binding_digest.as_str(),
-        )));
-    }
-    if decision.action != EndpointDecisionAction::CollectEvidence {
-        return Err(anyhow!("execution effect evidence is required"));
-    }
-    Ok(Some(stable_id(
-        "response_execution",
-        [
-            response_action_id.as_str(),
-            evidence_bundle_id.as_str(),
-            graph_content_hash,
-        ],
-    )))
-}
-
-fn response_execution_transition_id_prefix(status: &str) -> Option<&'static str> {
-    match status {
-        "failed" => Some("response_execution_failed"),
-        "partial" => Some("response_execution_partial"),
-        "rollback_pending" => Some("response_execution_rollback_pending"),
-        "rollback_failed" => Some("response_execution_rollback_failed"),
-        "expired" => Some("response_execution_expired"),
-        "cancelled" => Some("response_execution_cancelled"),
-        "rolled_back" => Some("response_execution_rolled_back"),
-        _ => None,
-    }
-}
-
-fn response_execution_reason_evidence_hash(evidence: &[EndpointReceiptEvidence]) -> Result<&str> {
-    let Some(reason) = evidence.iter().find(|item| item.key == "reason") else {
-        return Err(anyhow!("response transition reason evidence is required"));
-    };
-    require_evidence_hash_not_empty(reason, "response transition reason evidence")?;
-    Ok(reason.value_hash.as_str())
-}
-
-#[derive(Serialize)]
-#[serde(rename_all = "camelCase")]
-struct ResponseExecutionEffectBindingEntry {
-    key: String,
-    value_hash: String,
-}
-
-fn response_execution_id_from_effects(
-    response_action_id: &str,
-    evidence_bundle_id: &str,
-    effects: &[EndpointResponseExecutionEffect],
-) -> Result<String> {
-    let effect_binding_digest = response_execution_effect_binding_digest_from_effects(effects)?
-        .ok_or_else(|| anyhow!("response execution effect binding requires at least one effect"))?;
-    Ok(response_execution_id_from_effect_digest(
-        response_action_id,
-        evidence_bundle_id,
-        effect_binding_digest.as_str(),
-    ))
-}
-
-fn response_execution_id_from_effect_digest(
-    response_action_id: &str,
-    evidence_bundle_id: &str,
-    effect_binding_digest: &str,
-) -> String {
-    stable_id(
-        "response_execution",
-        [
-            response_action_id,
-            evidence_bundle_id,
-            effect_binding_digest,
-        ],
-    )
-}
-
-fn response_execution_transition_id_from_reason_hash(
-    prefix: &str,
-    response_action_id: &str,
-    evidence_bundle_id: &str,
-    rollback_ref: &str,
-    reason_hash: &str,
-) -> String {
-    stable_id(
-        prefix,
-        [
-            response_action_id,
-            evidence_bundle_id,
-            rollback_ref,
-            reason_hash,
-        ],
-    )
-}
-
-pub(crate) fn response_execution_effect_binding_digest_from_effects(
-    effects: &[EndpointResponseExecutionEffect],
-) -> Result<Option<String>> {
-    let entries = effects
-        .iter()
-        .map(|effect| ResponseExecutionEffectBindingEntry {
-            key: format!("executionEffect:{}", effect.effect_id),
-            value_hash: sha256(response_effect_evidence_value(effect).as_bytes()).to_hex_prefixed(),
-        })
-        .collect();
-    response_execution_effect_binding_digest(entries)
-}
-
-fn response_execution_effect_binding_digest_from_evidence(
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<Option<String>> {
-    let entries = evidence
-        .iter()
-        .filter(|item| item.key.starts_with("executionEffect:"))
-        .map(|item| ResponseExecutionEffectBindingEntry {
-            key: item.key.clone(),
-            value_hash: item.value_hash.clone(),
-        })
-        .collect();
-    response_execution_effect_binding_digest(entries)
-}
-
-fn response_execution_effect_binding_digest(
-    mut entries: Vec<ResponseExecutionEffectBindingEntry>,
-) -> Result<Option<String>> {
-    if entries.is_empty() {
-        return Ok(None);
-    }
-    entries.sort_by(|left, right| left.key.cmp(&right.key));
-    let value =
-        serde_json::to_value(entries).context("serialize response execution effect binding")?;
-    let canonical =
-        canonicalize_json(&value).context("canonicalize response execution effect binding")?;
-    Ok(Some(sha256(canonical.as_bytes()).to_hex_prefixed()))
-}
-
-pub(crate) fn response_rollback_id_from_effects(
-    execution_id: &str,
-    response_action_id: &str,
-    rollback_ref: &str,
-    effects: &[EndpointResponseExecutionEffect],
-) -> Result<String> {
-    let execution_id_hash = sha256(execution_id.as_bytes()).to_hex_prefixed();
-    let effect_binding_digest = response_rollback_effect_binding_digest_from_effects(effects)?
-        .ok_or_else(|| anyhow!("response rollback effect binding requires at least one effect"))?;
-    Ok(response_rollback_id_from_effect_digest(
-        response_action_id,
-        rollback_ref,
-        execution_id_hash.as_str(),
-        effect_binding_digest.as_str(),
-    ))
-}
-
-pub(crate) fn response_rollback_id_from_signed_evidence(
-    evidence: &[EndpointReceiptEvidence],
-    response_action_id: &str,
-    rollback_ref: &str,
-) -> Result<String> {
-    let Some(execution_id) = evidence.iter().find(|item| item.key == "executionId") else {
-        return Err(anyhow!("rollback execution id evidence is required"));
-    };
-    require_evidence_hash_not_empty(execution_id, "rollback execution id evidence")?;
-    let effect_binding_digest = response_rollback_effect_binding_digest_from_evidence(evidence)?
-        .ok_or_else(|| anyhow!("rollback effect evidence is required"))?;
-    Ok(response_rollback_id_from_effect_digest(
-        response_action_id,
-        rollback_ref,
-        execution_id.value_hash.as_str(),
-        effect_binding_digest.as_str(),
-    ))
-}
-
-fn response_rollback_id_from_effect_digest(
-    response_action_id: &str,
-    rollback_ref: &str,
-    execution_id_hash: &str,
-    effect_binding_digest: &str,
-) -> String {
-    stable_id(
-        "response_rollback",
-        [
-            response_action_id,
-            rollback_ref,
-            execution_id_hash,
-            effect_binding_digest,
-        ],
-    )
-}
-
-fn response_rollback_effect_binding_digest_from_effects(
-    effects: &[EndpointResponseExecutionEffect],
-) -> Result<Option<String>> {
-    let entries = effects
-        .iter()
-        .map(|effect| ResponseExecutionEffectBindingEntry {
-            key: format!("rollbackEffect:{}", effect.effect_id),
-            value_hash: sha256(response_effect_evidence_value(effect).as_bytes()).to_hex_prefixed(),
-        })
-        .collect();
-    response_execution_effect_binding_digest(entries)
-}
-
-fn response_rollback_effect_binding_digest_from_evidence(
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<Option<String>> {
-    let entries = evidence
-        .iter()
-        .filter(|item| item.key.starts_with("rollbackEffect:"))
-        .map(|item| ResponseExecutionEffectBindingEntry {
-            key: item.key.clone(),
-            value_hash: item.value_hash.clone(),
-        })
-        .collect();
-    response_execution_effect_binding_digest(entries)
-}
-
-pub(crate) fn response_effect_evidence_value(effect: &EndpointResponseExecutionEffect) -> String {
-    serde_json::to_value(effect)
-        .ok()
-        .and_then(|value| canonicalize_json(&value).ok())
-        .unwrap_or_else(|| effect.effect_id.clone())
-}
-
-pub(crate) fn response_acknowledgement_id_from_report_fields(
-    execution_id: &str,
-    response_action_id: &str,
-    rollback_ref: &str,
-    acknowledged_by: &str,
-    note: Option<&str>,
-    effects: &[EndpointResponseExecutionEffect],
-) -> String {
-    response_acknowledgement_id_from_report_fields_with_control(
-        execution_id,
-        response_action_id,
-        rollback_ref,
-        acknowledged_by,
-        note,
-        effects,
-        None,
-    )
-}
-
-pub(crate) fn response_acknowledgement_id_from_report_fields_with_control(
-    execution_id: &str,
-    response_action_id: &str,
-    rollback_ref: &str,
-    acknowledged_by: &str,
-    note: Option<&str>,
-    effects: &[EndpointResponseExecutionEffect],
-    control_correlation: Option<&EndpointResponseControlCorrelation>,
-) -> String {
-    let execution_id_hash = sha256(execution_id.as_bytes()).to_hex_prefixed();
-    let note_hash = note
-        .map(|note| sha256(note.as_bytes()).to_hex_prefixed())
-        .unwrap_or_else(response_acknowledgement_absent_note_marker);
-    let effect_binding_digest =
-        response_acknowledgement_effect_binding_digest_from_effects(effects)
-            .ok()
-            .flatten()
-            .unwrap_or_else(response_acknowledgement_absent_effect_marker);
-    let control_binding_digest = control_correlation.and_then(|control| {
-        response_control_acknowledgement_binding_digest_from_control(control).ok()
-    });
-    response_acknowledgement_id_from_evidence_hashes(
-        response_action_id,
-        rollback_ref,
-        execution_id_hash.as_str(),
-        acknowledged_by,
-        note_hash.as_str(),
-        effect_binding_digest.as_str(),
-        control_binding_digest.as_deref(),
-    )
-}
-
-pub(crate) fn response_acknowledgement_id_from_signed_evidence(
-    evidence: &[EndpointReceiptEvidence],
-    response_action_id: &str,
-    rollback_ref: &str,
-    acknowledged_by: &str,
-) -> Result<String> {
-    let Some(execution_id) = evidence.iter().find(|item| item.key == "executionId") else {
-        return Err(anyhow!("acknowledgement execution id evidence is required"));
-    };
-    require_evidence_hash_not_empty(execution_id, "acknowledgement execution id evidence")?;
-    let note_hash = if let Some(note) = evidence.iter().find(|item| item.key == "note") {
-        require_evidence_hash_not_empty(note, "acknowledgement note evidence")?;
-        note.value_hash.clone()
-    } else {
-        response_acknowledgement_absent_note_marker()
-    };
-    let effect_binding_digest =
-        response_acknowledgement_effect_binding_digest_from_evidence(evidence)?
-            .unwrap_or_else(response_acknowledgement_absent_effect_marker);
-    let control_binding_digest =
-        response_control_acknowledgement_binding_digest_from_evidence(evidence)?;
-    Ok(response_acknowledgement_id_from_evidence_hashes(
-        response_action_id,
-        rollback_ref,
-        execution_id.value_hash.as_str(),
-        acknowledged_by,
-        note_hash.as_str(),
-        effect_binding_digest.as_str(),
-        control_binding_digest.as_deref(),
-    ))
-}
-
-fn response_acknowledgement_id_from_evidence_hashes(
-    response_action_id: &str,
-    rollback_ref: &str,
-    execution_id_hash: &str,
-    acknowledged_by: &str,
-    note_hash: &str,
-    effect_binding_digest: &str,
-    control_binding_digest: Option<&str>,
-) -> String {
-    let mut parts = vec![
-        response_action_id,
-        rollback_ref,
-        execution_id_hash,
-        acknowledged_by,
-        note_hash,
-        effect_binding_digest,
-    ];
-    if let Some(control_binding_digest) = control_binding_digest {
-        parts.push(control_binding_digest);
-    }
-    stable_id("response_acknowledgement", parts)
-}
-
-fn response_acknowledgement_absent_note_marker() -> String {
-    "note:absent".to_string()
-}
-
-fn response_acknowledgement_absent_effect_marker() -> String {
-    "effect:absent".to_string()
-}
-
-fn response_acknowledgement_effect_binding_digest_from_effects(
-    effects: &[EndpointResponseExecutionEffect],
-) -> Result<Option<String>> {
-    let entries = effects
-        .iter()
-        .map(|effect| ResponseExecutionEffectBindingEntry {
-            key: format!("acknowledgementEffect:{}", effect.effect_id),
-            value_hash: sha256(response_effect_evidence_value(effect).as_bytes()).to_hex_prefixed(),
-        })
-        .collect();
-    response_execution_effect_binding_digest(entries)
-}
-
-fn response_acknowledgement_effect_binding_digest_from_evidence(
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<Option<String>> {
-    let entries = evidence
-        .iter()
-        .filter(|item| item.key.starts_with("acknowledgementEffect:"))
-        .map(|item| ResponseExecutionEffectBindingEntry {
-            key: item.key.clone(),
-            value_hash: item.value_hash.clone(),
-        })
-        .collect();
-    response_execution_effect_binding_digest(entries)
-}
-
-fn response_control_acknowledgement_binding_digest_from_control(
-    control: &EndpointResponseControlCorrelation,
-) -> Result<String> {
-    let mut entries = vec![
-        ResponseExecutionEffectBindingEntry {
-            key: "controlResponseActionId".to_string(),
-            value_hash: sha256(control.response_action_id.as_bytes()).to_hex_prefixed(),
-        },
-        ResponseExecutionEffectBindingEntry {
-            key: "controlTargetKind".to_string(),
-            value_hash: sha256(control.target_kind.as_bytes()).to_hex_prefixed(),
-        },
-        ResponseExecutionEffectBindingEntry {
-            key: "controlTargetId".to_string(),
-            value_hash: sha256(control.target_id.as_bytes()).to_hex_prefixed(),
-        },
-        ResponseExecutionEffectBindingEntry {
-            key: "controlAckTokenHash".to_string(),
-            value_hash: control.ack_token_hash.clone(),
-        },
-        ResponseExecutionEffectBindingEntry {
-            key: "controlAckStatus".to_string(),
-            value_hash: sha256(control.ack_status.as_bytes()).to_hex_prefixed(),
-        },
-    ];
-    if let Some(delivery_id) = &control.delivery_id {
-        entries.push(ResponseExecutionEffectBindingEntry {
-            key: "controlDeliveryId".to_string(),
-            value_hash: sha256(delivery_id.as_bytes()).to_hex_prefixed(),
-        });
-    }
-    if let Some(resulting_state) = &control.resulting_state {
-        entries.push(ResponseExecutionEffectBindingEntry {
-            key: "controlResultingState".to_string(),
-            value_hash: sha256(resulting_state.as_bytes()).to_hex_prefixed(),
-        });
-    }
-    response_execution_effect_binding_digest(entries)?
-        .ok_or_else(|| anyhow!("control acknowledgement binding evidence is required"))
-}
-
-fn response_control_acknowledgement_binding_digest_from_evidence(
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<Option<String>> {
-    if !evidence.iter().any(|item| item.key.starts_with("control")) {
-        return Ok(None);
-    }
-
-    let mut entries = Vec::new();
-    for (key, field_name) in [
-        (
-            "controlResponseActionId",
-            "control acknowledgement response action id evidence",
-        ),
-        (
-            "controlTargetKind",
-            "control acknowledgement target kind evidence",
-        ),
-        (
-            "controlTargetId",
-            "control acknowledgement target id evidence",
-        ),
-        (
-            "controlAckTokenHash",
-            "control acknowledgement token hash evidence",
-        ),
-        (
-            "controlAckStatus",
-            "control acknowledgement status evidence",
-        ),
-    ] {
-        let item = evidence
-            .iter()
-            .find(|item| item.key == key)
-            .ok_or_else(|| anyhow!("{field_name} is required"))?;
-        require_evidence_hash_not_empty(item, field_name)?;
-        entries.push(ResponseExecutionEffectBindingEntry {
-            key: key.to_string(),
-            value_hash: item.value_hash.clone(),
-        });
-    }
-    for (key, field_name) in [
-        (
-            "controlDeliveryId",
-            "control acknowledgement delivery id evidence",
-        ),
-        (
-            "controlResultingState",
-            "control acknowledgement resulting state evidence",
-        ),
-    ] {
-        if let Some(item) = evidence.iter().find(|item| item.key == key) {
-            require_evidence_hash_not_empty(item, field_name)?;
-            entries.push(ResponseExecutionEffectBindingEntry {
-                key: key.to_string(),
-                value_hash: item.value_hash.clone(),
-            });
-        }
-    }
-    Ok(Some(
-        response_execution_effect_binding_digest(entries)?
-            .ok_or_else(|| anyhow!("control acknowledgement binding evidence is required"))?,
-    ))
-}
-
-fn response_execution_bundle_id_from_signed_fields(
-    decision: &EndpointDecisionRecord,
-    graph_slice_id: &str,
-    graph_content_hash: &str,
-) -> Result<String> {
-    let rollback_ref = decision
-        .rollback_ref
-        .as_deref()
-        .ok_or_else(|| anyhow!("response rollback ref is required"))?;
-    let response_action_id = response_action_id_from_rollback_ref(&decision.action, rollback_ref)?;
-    let status = response_execution_status_from_decision(decision)?;
-    if status == "failed" {
-        return Ok(stable_id(
-            "evidence_bundle",
-            [
-                response_action_id.as_str(),
-                graph_slice_id,
-                graph_content_hash,
-                "failed",
-            ],
-        ));
-    }
-    Ok(stable_id(
-        "evidence_bundle",
-        [
-            response_action_id.as_str(),
-            graph_slice_id,
-            graph_content_hash,
-        ],
-    ))
-}
-
-fn require_response_execution_dry_run_evidence(evidence: &[EndpointReceiptEvidence]) -> Result<()> {
-    require_evidence_value_hash(evidence, "dryRun", "false", "execution dry-run evidence")
-}
-
-fn require_response_execution_actor_evidence(evidence: &[EndpointReceiptEvidence]) -> Result<()> {
-    let actor_hash = evidence_value_hash(evidence, "actorHash", "response actor evidence")?;
-    let execution_actor_hash =
-        evidence_value_hash(evidence, "executionActorHash", "execution actor evidence")?;
-    if !hex_strings_match(actor_hash, execution_actor_hash) {
-        return Err(anyhow!(
-            "execution actor evidence hash must match response actor evidence"
-        ));
-    }
-    Ok(())
-}
-
-fn require_response_rollback_execution_id_evidence(
-    evidence: &[EndpointReceiptEvidence],
-    signed_rollback_id: Option<&str>,
-    response_action_id: &str,
-    rollback_ref: &str,
-) -> Result<()> {
-    require_nonempty_hashed_evidence(evidence, "executionId", "rollback execution id evidence")?;
-    let rollback_id =
-        response_rollback_id_from_signed_evidence(evidence, response_action_id, rollback_ref)?;
-    if signed_rollback_id != Some(rollback_id.as_str()) {
-        return Err(anyhow!(
-            "rollback execution id evidence hash must match signed rollback proof"
-        ));
-    }
-    Ok(())
-}
-
-fn require_response_acknowledgement_execution_id_evidence(
-    evidence: &[EndpointReceiptEvidence],
-    signed_acknowledgement_id: Option<&str>,
-    response_action_id: &str,
-    rollback_ref: &str,
-    acknowledged_by: &str,
-) -> Result<()> {
-    require_nonempty_hashed_evidence(
-        evidence,
-        "executionId",
-        "acknowledgement execution id evidence",
-    )?;
-    let acknowledgement_id = response_acknowledgement_id_from_signed_evidence(
-        evidence,
-        response_action_id,
-        rollback_ref,
-        acknowledged_by,
-    )?;
-    if signed_acknowledgement_id != Some(acknowledgement_id.as_str()) {
-        return Err(anyhow!(
-            "acknowledgement execution id evidence hash must match signed acknowledgement proof"
-        ));
-    }
-    Ok(())
-}
-
-fn require_response_acknowledgement_note_evidence(
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<()> {
-    if evidence.iter().any(|item| item.key == "note") {
-        require_nonempty_hashed_evidence(evidence, "note", "acknowledgement note evidence")?;
-    }
-    Ok(())
-}
-
-fn require_response_effect_count_evidence(
-    evidence: &[EndpointReceiptEvidence],
-    effect_key_prefix: &str,
-    field_name: &str,
-) -> Result<()> {
-    let effect_count = evidence
-        .iter()
-        .filter(|item| item.key.starts_with(effect_key_prefix))
-        .count();
-    require_evidence_value_hash(
-        evidence,
-        "effectCount",
-        effect_count.to_string(),
-        field_name,
-    )
-}
-
-fn require_response_effect_evidence_hashes(
-    evidence: &[EndpointReceiptEvidence],
-    effect_key_prefix: &str,
-    field_name: &str,
-) -> Result<()> {
-    for effect_evidence in evidence
-        .iter()
-        .filter(|item| item.key.starts_with(effect_key_prefix))
-    {
-        require_evidence_hash_not_empty(effect_evidence, field_name)?;
-    }
-    Ok(())
-}
-
-fn require_response_execution_effect_type_evidence(
-    decision: &EndpointDecisionRecord,
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<()> {
-    let expected_effect_type = response_execution_effect_type_for_action(&decision.action);
-    require_response_typed_effect_evidence(
-        evidence,
-        "executionEffect:",
-        "executionEffectType:",
-        expected_effect_type,
-        "execution effect type evidence",
-    )
-}
-
-fn require_response_rollback_effect_type_evidence(
-    decision: &EndpointDecisionRecord,
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<()> {
-    let expected_effect_type = response_rollback_effect_type_for_action(&decision.action);
-    require_response_typed_effect_evidence(
-        evidence,
-        "rollbackEffect:",
-        "rollbackEffectType:",
-        expected_effect_type,
-        "rollback effect type evidence",
-    )
-}
-
-fn require_response_acknowledgement_effect_type_evidence(
-    decision: &EndpointDecisionRecord,
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<()> {
-    let expected_effect_type = response_execution_effect_type_for_action(&decision.action);
-    require_response_typed_effect_evidence(
-        evidence,
-        "acknowledgementEffect:",
-        "acknowledgementEffectType:",
-        expected_effect_type,
-        "acknowledgement effect type evidence",
-    )
-}
-
-fn require_response_typed_effect_evidence(
-    evidence: &[EndpointReceiptEvidence],
-    effect_key_prefix: &str,
-    effect_type_key_prefix: &str,
-    expected_effect_type: Option<&str>,
-    field_name: &str,
-) -> Result<()> {
-    let effect_ids = evidence
-        .iter()
-        .filter_map(|item| item.key.strip_prefix(effect_key_prefix))
-        .map(ToOwned::to_owned)
-        .collect::<BTreeSet<_>>();
-    let effect_type_ids = evidence
-        .iter()
-        .filter_map(|item| item.key.strip_prefix(effect_type_key_prefix))
-        .map(ToOwned::to_owned)
-        .collect::<BTreeSet<_>>();
-
-    if effect_ids != effect_type_ids {
-        return Err(anyhow!("{field_name} must match response effect evidence"));
-    }
-    if effect_ids.is_empty() {
-        return Ok(());
-    }
-
-    let expected_effect_type =
-        expected_effect_type.ok_or_else(|| anyhow!("{field_name} is invalid for action"))?;
-    for effect_id in effect_ids {
-        let effect_type_key = format!("{effect_type_key_prefix}{effect_id}");
-        require_evidence_value_hash(
-            evidence,
-            effect_type_key.as_str(),
-            expected_effect_type,
-            field_name,
-        )?;
-    }
-    Ok(())
-}
-
-fn response_execution_effect_type_for_action(
-    action: &EndpointDecisionAction,
-) -> Option<&'static str> {
-    match action {
-        EndpointDecisionAction::RestrictEgress => Some("restrict_egress"),
-        EndpointDecisionAction::QuarantineFile => Some("quarantine_file"),
-        EndpointDecisionAction::DisablePersistence => Some("disable_persistence"),
-        EndpointDecisionAction::RevokeGrant => Some("revoke_grant"),
-        EndpointDecisionAction::SuspendProcessTree => Some("suspend_process_tree"),
-        EndpointDecisionAction::TerminateProcessTree => Some("terminate_process_tree"),
-        EndpointDecisionAction::CollectEvidence => None,
-        _ => None,
-    }
-}
-
-fn response_rollback_effect_type_for_action(
-    action: &EndpointDecisionAction,
-) -> Option<&'static str> {
-    match action {
-        EndpointDecisionAction::RestrictEgress => Some("restore_egress"),
-        EndpointDecisionAction::QuarantineFile => Some("restore_quarantine_file"),
-        EndpointDecisionAction::DisablePersistence => Some("restore_persistence_file"),
-        EndpointDecisionAction::SuspendProcessTree => Some("resume_process_tree"),
-        _ => None,
-    }
-}
-
-fn require_response_acknowledgement_effect_evidence(
-    decision: &EndpointDecisionRecord,
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<()> {
-    let status = response_acknowledgement_status_from_decision(decision)?;
-    let effect_count = evidence
-        .iter()
-        .filter(|item| item.key.starts_with("acknowledgementEffect:"))
-        .count();
-    if matches!(decision.action, EndpointDecisionAction::CollectEvidence) && effect_count > 0 {
-        return Err(anyhow!(
-            "collect evidence acknowledgement effect evidence is invalid"
-        ));
-    }
-    if status == "succeeded"
-        && !matches!(decision.action, EndpointDecisionAction::CollectEvidence)
-        && effect_count == 0
-    {
-        return Err(anyhow!("acknowledgement effect evidence is required"));
-    }
-    Ok(())
-}
-
-fn require_response_execution_status_evidence(
-    decision: &EndpointDecisionRecord,
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<()> {
-    let expected_status = response_execution_status_from_decision(decision)?;
-    require_evidence_value_hash(
-        evidence,
-        "executionStatus",
-        expected_status,
-        "execution status evidence",
-    )
-}
-
-fn response_execution_status_from_decision(
-    decision: &EndpointDecisionRecord,
-) -> Result<&'static str> {
-    let Some(title) = decision.title.as_deref() else {
-        return Err(anyhow!("execution status title is required"));
-    };
-    let expected = match title {
-        "Endpoint response action executed" => "succeeded",
-        "Endpoint response action failed" => "failed",
-        "Endpoint response action partially executed" => "partial",
-        "Endpoint response rollback pending" => "rollback_pending",
-        "Endpoint response rollback failed" => "rollback_failed",
-        "Endpoint response action expired" => "expired",
-        "Endpoint response action cancelled" => "cancelled",
-        "Endpoint response action rolled back" => "rolled_back",
-        _ => return Err(anyhow!("execution status title is invalid")),
-    };
-    let expected_passed = expected == "succeeded";
-    if decision.passed != expected_passed {
-        return Err(anyhow!("execution status passed flag is inconsistent"));
-    }
-    Ok(expected)
-}
-
-fn require_response_rollback_status_evidence(
-    decision: &EndpointDecisionRecord,
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<()> {
-    if decision.title.as_deref() != Some("Endpoint response rollback executed") {
-        return Err(anyhow!("rollback status title is invalid"));
-    }
-    if !decision.passed {
-        return Err(anyhow!("rollback status passed flag is inconsistent"));
-    }
-    require_evidence_value_hash(
-        evidence,
-        "rollbackStatus",
-        "succeeded",
-        "rollback status evidence",
-    )
-}
-
-fn require_response_acknowledgement_status_evidence(
-    decision: &EndpointDecisionRecord,
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<()> {
-    let expected_status = response_acknowledgement_status_from_decision(decision)?;
-    require_evidence_value_hash(
-        evidence,
-        "acknowledgedStatus",
-        expected_status,
-        "acknowledgement status evidence",
-    )
-}
-
-fn response_acknowledgement_status_from_decision(
-    decision: &EndpointDecisionRecord,
-) -> Result<&str> {
-    let Some(title) = decision.title.as_deref() else {
-        return Err(anyhow!("acknowledgement status title is required"));
-    };
-    let Some(status) = title.strip_prefix("Endpoint response execution acknowledged: ") else {
-        return Err(anyhow!("acknowledgement status title is invalid"));
-    };
-    match status {
-        "succeeded" | "failed" | "partial" | "expired" | "cancelled" | "rolled_back" => {}
-        _ => return Err(anyhow!("acknowledgement status title is invalid")),
-    }
-    if !decision.passed {
-        return Err(anyhow!(
-            "acknowledgement status passed flag is inconsistent"
-        ));
-    }
-    Ok(status)
-}
-
-fn require_response_acknowledged_by_evidence(
-    actor: &EndpointDecisionActor,
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<()> {
-    let acknowledged_by = actor
-        .agent_id
-        .as_deref()
-        .ok_or_else(|| anyhow!("acknowledged-by actor identity is required"))?;
-    require_nonempty(acknowledged_by, "acknowledged-by actor identity")?;
-    require_evidence_value_hash(
-        evidence,
-        "acknowledgedBy",
-        acknowledged_by,
-        "acknowledged-by evidence",
-    )
-}
-
-fn require_response_control_acknowledgement_evidence(
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<()> {
-    let has_control_evidence = evidence.iter().any(|item| item.key.starts_with("control"));
-    if !has_control_evidence {
-        return Ok(());
-    }
-
-    require_nonempty_hashed_evidence(
-        evidence,
-        "controlResponseActionId",
-        "control acknowledgement response action id evidence",
-    )?;
-    require_control_target_kind_evidence(evidence)?;
-    require_nonempty_hashed_evidence(
-        evidence,
-        "controlTargetId",
-        "control acknowledgement target id evidence",
-    )?;
-    require_control_ack_token_hash_evidence(evidence)?;
-    require_control_ack_status_evidence(evidence)?;
-
-    for (key, field_name) in [
-        (
-            "controlDeliveryId",
-            "control acknowledgement delivery id evidence",
-        ),
-        (
-            "controlResultingState",
-            "control acknowledgement resulting state evidence",
-        ),
-    ] {
-        if evidence.iter().any(|item| item.key == key) {
-            require_nonempty_hashed_evidence(evidence, key, field_name)?;
-        }
-    }
-
-    Ok(())
-}
-
-fn require_control_ack_token_hash_evidence(evidence: &[EndpointReceiptEvidence]) -> Result<()> {
-    let Some(item) = evidence
-        .iter()
-        .find(|item| item.key == "controlAckTokenHash")
-    else {
-        return Err(anyhow!(
-            "control acknowledgement token hash evidence is required"
-        ));
-    };
-    if item.redaction_class != EndpointEvidenceRedactionClass::HashOnly || item.raw_value.is_some()
-    {
-        return Err(anyhow!(
-            "control acknowledgement token hash evidence must be hash-only"
-        ));
-    }
-    require_evidence_hash_not_empty(item, "control acknowledgement token hash evidence")
-}
-
-fn require_control_ack_status_evidence(evidence: &[EndpointReceiptEvidence]) -> Result<()> {
-    let Some(item) = evidence.iter().find(|item| item.key == "controlAckStatus") else {
-        return Err(anyhow!(
-            "control acknowledgement status evidence is required"
-        ));
-    };
-    require_evidence_hash_not_empty(item, "control acknowledgement status evidence")?;
-    let allowed_statuses = [
-        "acknowledged",
-        "rejected",
-        "failed",
-        "expired",
-        "rolled_back",
-    ];
-    if allowed_statuses.iter().any(|status| {
-        let expected_hash = sha256(status.as_bytes()).to_hex_prefixed();
-        hex_strings_match(expected_hash.as_str(), item.value_hash.as_str())
-    }) {
-        return Ok(());
-    }
-    Err(anyhow!(
-        "control acknowledgement status evidence must be acknowledged, rejected, failed, expired, or rolled_back"
-    ))
-}
-
-fn require_control_target_kind_evidence(evidence: &[EndpointReceiptEvidence]) -> Result<()> {
-    let Some(item) = evidence.iter().find(|item| item.key == "controlTargetKind") else {
-        return Err(anyhow!(
-            "control acknowledgement target kind evidence is required"
-        ));
-    };
-    require_evidence_hash_not_empty(item, "control acknowledgement target kind evidence")?;
-    let allowed_target_kinds = [
-        "endpoint",
-        "runtime",
-        "session",
-        "principal",
-        "grant",
-        "swarm",
-        "project",
-    ];
-    if allowed_target_kinds.iter().any(|target_kind| {
-        let expected_hash = sha256(target_kind.as_bytes()).to_hex_prefixed();
-        hex_strings_match(expected_hash.as_str(), item.value_hash.as_str())
-    }) {
-        return Ok(());
-    }
-    Err(anyhow!(
-        "control acknowledgement target kind evidence must be endpoint, runtime, session, principal, grant, swarm, or project"
-    ))
-}
-
-fn require_response_action(action: &EndpointDecisionAction) -> Result<()> {
-    if matches!(
-        action,
-        EndpointDecisionAction::RestrictEgress
-            | EndpointDecisionAction::SuspendProcessTree
-            | EndpointDecisionAction::TerminateProcessTree
-            | EndpointDecisionAction::QuarantineFile
-            | EndpointDecisionAction::RevokeGrant
-            | EndpointDecisionAction::DisablePersistence
-            | EndpointDecisionAction::CollectEvidence
-    ) {
-        return Ok(());
-    }
-
-    Err(anyhow!(
-        "response action must be a bounded local response action"
-    ))
-}
-
-fn require_response_action_for_family(
-    family: &EndpointDecisionReceiptFamily,
-    decision: &EndpointDecisionRecord,
-) -> Result<()> {
-    require_response_action(&decision.action)?;
-    if decision.action != EndpointDecisionAction::TerminateProcessTree {
-        return Ok(());
-    }
-
-    if family == &EndpointDecisionReceiptFamily::ResponseRequest
-        && response_request_dry_run_from_decision(decision)?
-    {
-        return Ok(());
-    }
-
-    Err(anyhow!(
-        "terminate_process_tree response proofs are limited to dry-run response requests"
-    ))
-}
-
-fn require_response_actor_context(actor: &EndpointDecisionActor) -> Result<()> {
-    let has_actor_context = [
-        actor.user_id.as_deref(),
-        actor.session_id.as_deref(),
-        actor.agent_id.as_deref(),
-        actor.workload_id.as_deref(),
-        actor.approval_id.as_deref(),
-    ]
-    .into_iter()
-    .flatten()
-    .any(|value| !value.trim().is_empty());
-
-    if has_actor_context {
-        return Ok(());
-    }
-    Err(anyhow!("response actor context is required"))
-}
-
-fn require_response_actor_evidence(
-    actor: &EndpointDecisionActor,
-    evidence: &[EndpointReceiptEvidence],
-) -> Result<()> {
-    require_evidence_value_hash(
-        evidence,
-        "actorHash",
-        endpoint_decision_actor_content_hash(actor),
-        "response actor evidence",
-    )
-}
